Module Ascii_table

include module type of struct include Ascii_table_kernel end

module Attr = Ascii_table_kernel.Attr

module Align = Ascii_table_kernel.Align

module Display = Ascii_table_kernel.Display

module Column = Ascii_table_kernel.Column

module Screen = Ascii_table_kernel.Screen

val draw : ?display:Display.t -> ?spacing:int -> ?limit_width_to:int -> ?header_attr:Attr.t list -> ?display_empty_rows:bool -> prefer_split_on_spaces:bool -> 'row Column.t list -> 'row list -> Screen.t option

module Table_char = Ascii_table_kernel.Table_char

type ('row, 'rest) renderer = ?display:Display.t -> ?spacing:int -> ?limit_width_to:int -> ?header_attr:Attr.t list -> ?bars:[ `Ascii | `Unicode ] -> ?display_empty_rows:bool -> ?prefer_split_on_spaces:bool -> 'row Column.t list -> 'row list -> 'rest

val output : ( _, oc:Core.Out_channel.t -> unit ) renderer

The idea is that you have a Column.t list and a list of rows, where each row contains the data for each column. So e.g. 'a could be a record type col_1 : int; col_2 : string, where the first column pulls out col_1 and the second column pulls out col_2. *

val to_string_noattr : ( _, string ) renderer

val to_string : ( _, string ) renderer

type 'rest simple_list := ?index:bool -> ?display:Display.t -> ?spacing:int -> ?limit_width_to:int -> ?header_attr:Attr.t list -> ?bars:[ `Ascii | `Unicode ] -> ?display_empty_rows:bool -> ?prefer_split_on_spaces:bool -> string list -> string list list -> 'rest

simple_list has the same defaults as renderer unless specified

val simple_list_table : ?oc:Core.Out_channel.t -> unit simple_list

val simple_list_table_string : string simple_list