Module Ascii_table_kernel.Column

module Align : sig ... end

module Show : sig ... end

type 'a t

val sexp_of_t : ( 'a -> Sexplib0.Sexp.t ) -> 'a t -> Sexplib0.Sexp.t

val create : ?align:Align.t -> ?min_width:int -> ?max_width:int -> ?show:[ `Yes | `No | `If_not_empty ] -> string -> ( 'a -> string ) -> 'a t

creates a column given the header and the to-string function

val create_attr : ?align:Align.t -> ?min_width:int -> ?max_width:int -> ?show:[ `Yes | `No | `If_not_empty ] -> string -> ( 'a -> Ansi_kernel.Attr.t list * string ) -> 'a t

like create, except that the to_string function must provide a list of attributes.

val lift : 'a t -> f:( 'b -> 'a ) -> 'b t

val align : _ t -> Align.t

val header : 'a t -> string

val show : _ t -> Show.t

val to_data : 'a t -> 'a -> Ansi_kernel.Attr.t list * string list

module Of_field : sig ... end

This module is used for constructing lists of 'a ts from a record's fields. The intention is to use Fields.to_list to obtain a list.