Module Ascii_table_kernel.Screen

type t

A Screen.t represents a table after all of the layout calculations have been done.

val render : t -> bars:[ `Ascii | `Unicode ] -> output:( Attr.t list -> Core.Buffer.t -> unit ) -> close:( Core.Buffer.t -> 'a ) -> 'a

val to_string : t -> bars:[ `Ascii | `Unicode ] -> string_with_attr:( Attr.t list -> string -> string ) -> string

Given a way to annotate strings with attributes, a t can be output to a string.