Module type Console.Io

type 'a t

type 'a fmt

type out_channel

include Core.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:( 'a -> 'b t ) -> 'b t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

map t ~f is t >>| f.

val join : 'a t t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> unit t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad:

val output_string : out_channel -> string -> unit

val output : out_channel -> buf:bytes -> pos:int -> len:int -> unit

val stderr : out_channel

val stdout : out_channel

val flush : out_channel -> unit t

val print_string : string -> unit

val fprintf : attrs:string -> out_channel -> 'a fmt -> 'a

val fold_left : 'a Core.List.t -> init:'b -> f:( 'b -> 'a -> 'b t ) -> 'b t

val stdout_isatty : unit -> bool t

val capable : unit -> bool t